🔧 Add getBounties endpoint to UserController
Added a new static method `getBounties` to the `UserController` class in `user.controller.js`. This method retrieves bounties associated with a user by their ID. The retrieved bounties are then returned as a response along with a success message. If an error occurs during the retrieval process, an error message is returned instead.

In `users.js`, added a new route `/users/:id/bounties` that maps to the `getBounties` method in the `UserController`.

No other changes were made in this commit.